Bugfix: Fixing energy input on reprocessors
Reprocessors no longer give 100MJ per 1MJ input. UE conversion is now
correct too.

+++ erogenousbeef/bigreactors/common/multiblock/MultiblockReactor.java
@@ -489,12 +489,12 @@ public class MultiblockReactor extends MultiblockControllerBase implements IBeef
 
 	// IBeefPowerStorage
 	@Override
-	public int getEnergyStored() {
-		return (int)storedEnergy;
+	public float getEnergyStored() {
+		return (float)storedEnergy;
 	}
 
 	@Override
-	public int getMaxEnergyStored() {
+	public float getMaxEnergyStored() {
 		return maxEnergyStored;
 	}
